cloud/backend/requirement.txt
tensorflow
flask
gcsfs
h5py
numpy
scikit-learn
pandas
mysql-connector-python